created user dialog widget
