+++ lib/common/constants/strings.dart
@@ -12,12 +12,12 @@ const String noInternetErrorMessage =
     'No internet connection. Please check your connection and try again.';
 const String editUserDialogTitle = 'Edit User';
 const String addUserDialogTitle = 'Add User';
-const String editUserDialogNameTextFieldTitle = 'Name';
-const String editUserDialogEmailTextFieldTitle = 'Email';
-const String editUserDialogGenderLabel = 'Gender';
-const String editUserDialogStatusLabel = 'Status';
-const String editUserDialogCancelButtonText = 'Cancel';
-const String editUserDialogSaveAndCloseButtonText = 'Save & Close';
+const String userDialogNameTextFieldTitle = 'Name';
+const String userDialogEmailTextFieldTitle = 'Email';
+const String userDialogGenderLabel = 'Gender';
+const String userDialogStatusLabel = 'Status';
+const String userDialogCancelButtonText = 'Cancel';
+const String userDialogSaveAndCloseButtonText = 'Save & Close';
 const String maleGender = 'male';
 const String femaleGender = 'female';
 const String activeStatus = 'active';
