+++ lib/features/user/add/ui/add_user_dialog.dart
@@ -1,11 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:user_app/common/widgets/user_dialog.dart';
 import 'package:user_app/features/user/list/network/user_api.dart';
-import 'package:user_app/features/user/list/network/user_response.dart';
 import '../../../../common/constants/enums.dart';
 import '../../../../common/constants/strings.dart';
 import '../../../../common/state/ui_state.dart';
-import '../../../../common/utils/utils.dart';
 import '../../list/bloc/user_cubit.dart';
 import '../bloc/add_user_cubit.dart';
 
@@ -25,22 +24,6 @@ class AddUserDialogState extends State<AddUserDialog> {
   late String email = '';
   late TextEditingController nameController = TextEditingController(text: '');
   late TextEditingController emailController = TextEditingController(text: '');
-  bool _isEmailValid = true;
-  bool _isNameValid = true;
-
-  void _onEmailChanged(String value) {
-    name = value;
-    setState(() {
-      _isEmailValid = (validateEmail(value) == null);
-    });
-  }
-
-  void _onNameChanged(String value) {
-    email = value;
-    setState(() {
-      _isNameValid = validateName(value) == null;
-    });
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -48,123 +31,49 @@ class AddUserDialogState extends State<AddUserDialog> {
       create: (context) => AddUserCubit(UserApi()),
       child: BlocConsumer<AddUserCubit, UiState<bool>>(
         builder: (context, state) {
-          if (state is Default || state is Loading || state is Error) {
-            return AlertDialog(
-              title: Text(
-                addUserDialogTitle,
-              ),
-              backgroundColor: Colors.white,
-              content: SizedBox(
-                width: 300,
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    _AddUserDialogTextField(
-                      controller: nameController,
-                      title: editUserDialogNameTextFieldTitle,
-                      keyboardType: TextInputType.name,
-                      onChanged: _onNameChanged,
-                    ),
-                    SizedBox(height: 10),
-                    _AddUserDialogTextField(
-                      controller: emailController,
-                      title: editUserDialogEmailTextFieldTitle,
-                      keyboardType: TextInputType.emailAddress,
-                      onChanged: _onEmailChanged,
-                    ),
-                    SizedBox(height: 15),
-                    _RadioListWidget(
-                      label: editUserDialogGenderLabel,
-                      values: [Gender.male, Gender.female],
-                      groupValue: gender,
-                      onChanged: (Gender? value) {
-                        if (value != null) {
-                          gender = value;
-                        }
-                        setState(
-                          () {
-                            gender = value!;
-                          },
-                        );
-                      },
-                    ),
-                    SizedBox(height: 10),
-                    _RadioListWidget(
-                      label: editUserDialogStatusLabel,
-                      values: [Status.active, Status.inactive],
-                      groupValue: status,
-                      onChanged: (Status? value) {
-                        if (value != null) status = value;
-                        setState(
-                          () {
-                            status = value!;
-                          },
-                        );
-                      },
-                    ),
-                  ],
-                ),
-              ),
-              actions: [
-                TextButton(
-                  onPressed: () {
-                    Navigator.of(context).pop();
-                  },
-                  child: Text(
-                    editUserDialogCancelButtonText,
-                  ),
-                ),
-                ElevatedButton(
-                  onPressed: _isEmailValid &&
-                          _isNameValid &&
-                          status != Status.undefined &&
-                          gender != Gender.undefined
-                      ? () async {
-                          name = nameController.text;
-                          email = emailController.text;
-                          gender = gender;
-                          status = status;
-
-                          await context.read<AddUserCubit>().addUser(
-                              UserResponse(
-                                  userName: name,
-                                  userGender: gender,
-                                  userEmail: email,
-                                  userStatus: status,
-                                  userId: 0));
-                        }
-                      : null,
-                  child: (state is Loading)
-                      ? CircularProgressIndicator()
-                      : Text(
-                          editUserDialogSaveAndCloseButtonText,
-                        ),
-                ),
-              ],
-            );
-          } else {
+          if (state is Success) {
             return SizedBox.shrink();
+          } else {
+            return UserDialog(
+              action: DialogAction.add,
+              state: state,
+            );
           }
         },
         listener: (context, state) {
-          if (state is Error) {
+          if (state is Success) {
+            Navigator.of(context).pop();
+            context.read<UserCubit>().fetchUsers();
+          }
+          if (state is Error || state is Success) {
             ScaffoldMessenger.of(context).showSnackBar(
               SnackBar(
                 content: Row(
                   children: [
-                    Icon(Icons.error_outline, color: Colors.white),
+                    Icon(
+                        (state is Error)
+                            ? Icons.error_outline
+                            : Icons.check_circle_outline,
+                        color: Colors.white),
                     SizedBox(width: 10),
                     Expanded(
                       child: Text(
-                        state.getError().toString(),
-                        style: TextStyle(color: Colors.white, fontSize: 16),
+                        (state is Error)
+                            ? unexpectedErrorMessage
+                            : userAddedSuccessMessage,
+                        style: TextStyle(
+                          color: Colors.white,
+                          fontSize: 16,
+                        ),
                         maxLines: 2,
                         overflow: TextOverflow.ellipsis,
                       ),
                     ),
                   ],
                 ),
-                backgroundColor: Colors.red.shade600,
+                backgroundColor: (state is Error)
+                    ? Colors.red.shade600
+                    : Colors.green.shade600,
                 behavior: SnackBarBehavior.floating,
                 margin: EdgeInsets.only(
                   bottom: MediaQuery.of(context).viewInsets.bottom + 16,
@@ -178,124 +87,8 @@ class AddUserDialogState extends State<AddUserDialog> {
               ),
             );
           }
-          if (state is Success) {
-            Navigator.of(context).pop();
-            context.read<UserCubit>().fetchUsers();
-            ScaffoldMessenger.of(context).showSnackBar(
-              SnackBar(
-                content: Row(
-                  children: [
-                    Icon(Icons.check_circle_outline, color: Colors.white),
-                    SizedBox(width: 10),
-                    Expanded(
-                      child: Text(
-                        userAddedSuccessMessage,
-                        style: TextStyle(color: Colors.white, fontSize: 16),
-                        maxLines: 2,
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                    ),
-                  ],
-                ),
-                backgroundColor: Colors.green.shade600,
-                behavior: SnackBarBehavior.floating,
-                margin: EdgeInsets.all(16),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(12),
-                ),
-                duration: Duration(seconds: 2),
-              ),
-            );
-          }
         },
       ),
     );
-  }
-}
-
-class _AddUserDialogTextField extends StatelessWidget {
-  final TextEditingController controller;
-  final String title;
-  final TextInputType keyboardType;
-  final void Function(String) onChanged;
-
-  const _AddUserDialogTextField({
-    required this.controller,
-    required this.title,
-    required this.keyboardType,
-    required this.onChanged,
-  });
-
-  String? _getTextFieldError(TextInputType keyboardType) {
-    if (keyboardType == TextInputType.emailAddress) {
-      return validateEmail(controller.text);
-    } else if (keyboardType == TextInputType.name) {
-      return validateName(controller.text);
-    }
-    return null;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return TextField(
-      controller: controller,
-      decoration: InputDecoration(
-        labelText: title,
-        errorText: _getTextFieldError(keyboardType),
-      ),
-      keyboardType: keyboardType,
-      onChanged: (value) {
-        onChanged(value);
-        (context as Element).markNeedsBuild();
-      },
-    );
-  }
-}
-
-class _RadioListWidget<T extends Enum> extends StatelessWidget {
-  final String label;
-  final List<T> values;
-  final T groupValue;
-  final ValueChanged<T?> onChanged;
-
-  const _RadioListWidget({
-    super.key,
-    required this.values,
-    required this.groupValue,
-    required this.label,
-    required this.onChanged,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(children: [
-      Align(
-        alignment: Alignment.centerLeft,
-        child: Text(
-          label,
-          style: TextStyle(
-            fontSize: 12,
-          ),
-        ),
-      ),
-      Row(
-        children: values.map((T g) {
-          return Expanded(
-            child: RadioListTile<T>(
-              title: Text(
-                g.name,
-                style: TextStyle(
-                  fontSize: 16,
-                ),
-              ),
-              contentPadding: EdgeInsets.zero,
-              value: g,
-              groupValue: groupValue,
-              onChanged: onChanged,
-            ),
-          );
-        }).toList(),
-      ),
-    ]);
   }
 }
